runtime: adjust frame pointer of first g0 frame for amd64
When a g switches to the g0 stack, the first frame of the g0 stack
contains a frame pointer that points to the frame on the g stack that
invoked it. Adjust this pointer on the g0 stack after adjusting the
frames of the g stack, but before calling stackfree. This ensures that
there is no moment in time where an async frame pointer unwinder could
unwind into a stack that has already been freed.

Also remove the code to restore sched.bp in systemstack. This operation
would be entirely redundant now. Even before it was mostly redundant
because systemstack's epilogue was also restoring a correctly adjusted
frame pointer shortly afterwards.

Add an assertion in systemstack that verifies that all of this is
working correctly. Without the change to copystack, this check throws
when running TestDeferHeapAndStack.

This patch has no impact on traceEvent because it starts unwinding at
gp.sched.bp which is always pointing to a g stack and therefore not
visiting any frame pointers on the g0 stack.

Alternatively we could just accept this issue and document it in
copystack. Async frame pointer unwinding is subject to race conditions
anyway. E.g. unwinding a leaf frame before it finishes its frame pointer
prologue causes its calling frame to be missed. The initial motivation
for this patch was to figure out if restoring sched.bp in systemstack
was redundant or not.

Also add a newline to the string in badsystemstack while we are here.
writeErrStr expects a trailing newline in order to produce good output,
and this is the only place that was missing it.

Updates #59692

Change-Id: Icf3266475d1439ba5f34d54bd8b76bdb72592654