cmd/api: make api/next/* entries required, not optional
We want the API check to catch if some API present in api/next/*
files is no longer implemented in the tree, and report it in the
same CL that is making the change (by failing loudly otherwise).
That helps confirm the API change is intentional.

Remove the log/slog.Record.Attrs entry from api/next/56345.txt,
it has been replaced with an updated API in api/next/59060.txt.

Change-Id: I73330dd5fd3f5706a1fdf13b2bf8e0f24c6b48e5