internal/fuzz: convert custom mutator panics to returned errors
If a fuzz worker process panics (or otherwise exits) due to an error
outside of the test function, the dev UX is suboptimal: the
coordinator assumes that an interesting case was discovered, and does
not print any error messages emitted by the worker (such as a panic
stack trace).

This change improves custom mutator UX by converting panics resulting
from failed calls into internal errors that are reported to the
coordinator.

Change-Id: Ie49c362d3e9eaab124eddab361eb8151a1003c34