internal/fuzz: convert internal worker panics to returned errors
If a fuzz worker process panics (or otherwise exits) due to an error
outside of the test function, the dev UX is suboptimal: the
coordinator assumes that an interesting case was discovered, and does
not print any error messages emitted by the worker (such as a panic
stack trace). Ideally the coordinator would treat such errors as the
internal errors that they are, and only treat errors inside the test
function as interesting. That change would require significant effort
and add considerable complexity, so this commit instead converts some
worker-terminating errors into errors reported to the coordinator.

I am not currently experiencing any of these errors, but this change
is a prerequisite to extending mutator.mutate to support custom types
that implement their own mutate and marshal/unmarshal methods. This
commit enables developer-friendly propagation of any errors that might
occur in those custom methods.

Change-Id: Ie49c362d3e9eaab124eddab361eb8151a1003c34