cmd/internal/obj/arm64: simplify oplook
Avoid continually adding and subtracting one from classes.

This code has its roots in the original 7l C code - the intent is to
ensure that the value of zero means uninitialised, such that once a
class is determined it can be cached. This is currently done by adding
one to the class, then subtracting it again on use.

Instead of doing this, simply offset the classes by one. This allows
for the same behaviour without the +/-1 dances.

Change-Id: Ib4383a73dc38ea823550e0adeb5aece2ab49bcc5