REBASE: cmd/compile: enable profile guided code specialization
[This is a rebased copy of CL 484838, patch set 4 to serve as a
temporary diffbase for the following CL.]

DO NOT SUBMIT
DO NOT REVIEW

Go v1.20 released PGO as a preview technology which used profile
information in inlining optimization. This code change adds a new code
specialization pass based on the profile information, i.e., a hot
interface method call is converted into an "if-then-else" code block
based on the hot dynamic type found in the profile.

The algorithm is described at
https://github.com/golang/proposal/blob/master/design/55022-pgo-implementation.md

Change-Id: Ief63b496c00177b145a40bcd08f7b534de192c16