cmd/compile: enable PGO-driven call specialization
DO NOT SUBMIT
DO NOT REVIEW

[This is based on CL 484838 by rajbarik@uber.com.]

This change adds a new code specialization pass based on the profile
information, i.e., a hot interface method call is converted into an
"if-then-else" code block based on the hot dynamic type found in the
profile.

This refines the approach from CL 484838. Notably:

* Indirect calls are integrated directly into Profile.WeightedCG rather
  than using separate side structures. Fixes several bugs along the way,
  mostly notably the old approach failed to match pointer recevier
  methods (pkg.(*Foo).Bar). Still needs more cleanup.
* The call node replacement is significantly simplified and generalized.

Fixes #59959

Change-Id: I8bedb516139695ee4069650b099d05957b7ce5ee