misc/cgo: move easy tests to cmd/cgo/testdata
This moves most misc/cgo tests to cmd/cgo/testdata. This is mostly a
trivial rename and updating dist/test.go for the new paths. A few
tests were sensitive to what path they ran in, so we update those.

This will let these tests access facilities in internal/testenv.

For #37486.

Change-Id: I3ed417c7c22d9b667f2767c0cb1f59118fcd4af6