cmd/cover: add new "emit meta file" mode for packages without tests
Introduce a new mode of execution for instrumenting packages that have
no test files. Instead of just skipping packages with no test files
(during "go test -cover" runs), the go command will invoke cmd/cover
on the package passing in an option in the config file indicating that
it should emit a coverage meta-data file directly for the package if
it has any functions. Note that this CL doesn't actually wire up
this functionality in the Go command, that will come in a later patch.

Updates #27261.
Updates #58770
Updates #24570.

Change-Id: I01e8a3edb62441698c7246596e4bacbd966591c3