cmd/link: suppress symbol table on Mach-O when -s
Currently, on Mach-O, we don't strip the symbol table even the -s
flag is set. This CL makes it suppress the symbol table, as
documented.

On Mach-O, even with -s, we still need to keep dynamic symbol
references. Otherwise the dynamic linker cannot resolve them and
the binary doesn't run. (Interestingly, for a PIE binary it is
okay to strip the symbol table entirely. We keep the dynamic
symbol references for consistency. And this is also in consistent
with what the system "strip" command does.)

Change-Id: I39c572553fe0215ae3bdf5349bf2bab7205fbdc9