test,internal/testdir: don't set GOOS/GOARCH
The test directory driver currently sets the GOOS/GOARCH environment
variables if they aren't set. This appears to be in service of a
single test, test/env.go, which was introduced in September 2008 along
with os.Getenv. It's not entirely clear what that test is even trying
to check, and at any rate we certainly have better tests of Getenv
now, so to simplify the test directory driver, we eliminate this test
and where the driver sets GOOS/GOARCH.

Change-Id: I5acc0093b557c95d1f0a526d031210256a68222d