runtime: add execution tracer v2 behind GOEXPERIMENT=exectracer2
This change mostly implements the design described in #60773 and
includes a new scalable parser for the new trace format, available in
internal/trace/v2. I'll leave this commit message short because this is
clearly an enormous CL with a lot of detail.

This change does not hook up the new tracer into cmd/trace yet. A
follow-up CL will handle that.

TODO: Needs a doc comment summary of the execution tracer's design and
the parser's algorithm.

For #60773.

Change-Id: I5d2aca2cc07580ed3c76a9813ac48ec96b157de0